Restructure the repository
Header files are now separate from source files.

**NOTE** This commit may break projects expecting header files in
a particular location.
