+++ CMakeLists.txt
@@ -2,5 +2,5 @@ cmake_minimum_required(VERSION 3.0)
 project(text-complex C)
 
 add_subdirectory(deps)
-add_subdirectory(tcmplx-access)
+add_subdirectory(src)
 add_subdirectory(tests)
