+++ include/text-complex/access/inscopy.h
