+++ src/ctxtmap.c
@@ -4,9 +4,9 @@
  * @author Cody Licorish (svgmovement@gmail.com)
  */
 #define TCMPLX_A_WIN32_DLL_INTERNAL
-#include "ctxtmap.h"
-#include "api.h"
-#include "util.h"
+#include "text-complex/access/ctxtmap.h"
+#include "text-complex/access/api.h"
+#include "text-complex/access/util.h"
 #include <string.h>
 
 struct tcmplxA_ctxtmap {
