+++ src/util.c
@@ -4,7 +4,7 @@
  * \author Cody Licorish (svgmovement@gmail.com)
  */
 #define TCMPLX_A_WIN32_DLL_INTERNAL
-#include "util.h"
+#include "text-complex/access/util.h"
 #include <stdlib.h>
 
 void* tcmplxA_util_malloc(size_t sz) {
