+++ tests/tca_api.c
@@ -1,5 +1,5 @@
 
-#include "../tcmplx-access/api.h"
+#include "text-complex/access/api.h"
 #include <stdio.h>
 
 int main(int argc, char **argv) {
