+++ tests/tca_util.c
@@ -1,9 +1,9 @@
 
-#include "../tcmplx-access/util.h"
+#include "text-complex/access/util.h"
+#include "munit/munit.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include "munit/munit.h"
 
 #ifndef NDEBUG
 #  define test_sharp2(s) #s
